refactor(type): P0.1 - Move the Node type from store to shared type folder.
